+++ src/Microsoft.WindowsPackageManager.Rest/Microsoft.WindowsPackageManager.Rest.csproj
@@ -46,7 +46,6 @@
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="Microsoft.Msix.Utils" Version="*" />
-    <PackageReference Include="Microsoft.Cloud.InstrumentationFramework.NetStd" Version="1.2.1011530001" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.11" />
   </ItemGroup>
